Refs #23647. Move rtps domain code into new folder.
