+++ src/cpp/fastdds/builtin/type_lookup_service/TypeLookupManager.cpp
@@ -40,7 +40,7 @@
 #include <rtps/builtin/discovery/participant/PDP.h>
 #include <rtps/participant/RTPSParticipantImpl.hpp>
 #include <rtps/reader/StatefulReader.hpp>
-#include <rtps/RTPSDomainImpl.hpp>
+#include <rtps/domain/RTPSDomainImpl.hpp>
 #include <rtps/writer/StatefulWriter.hpp>
 
 namespace eprosima {
