+++ src/cpp/fastdds/builtin/type_lookup_service/TypeLookupRequestListener.cpp
@@ -23,7 +23,7 @@
 
 #include <fastdds/builtin/type_lookup_service/TypeLookupManager.hpp>
 #include <rtps/participant/RTPSParticipantImpl.hpp>
-#include <rtps/RTPSDomainImpl.hpp>
+#include <rtps/domain/RTPSDomainImpl.hpp>
 #include <rtps/reader/StatefulReader.hpp>
 
 using eprosima::fastdds::rtps::RTPSReader;
