+++ src/cpp/fastdds/domain/DomainParticipantFactory.cpp
@@ -33,7 +33,7 @@
 #include <fastdds/domain/DomainParticipantImpl.hpp>
 #include <fastdds/log/LogResources.hpp>
 #include <rtps/history/TopicPayloadPoolRegistry.hpp>
-#include <rtps/RTPSDomainImpl.hpp>
+#include <rtps/domain/RTPSDomainImpl.hpp>
 #include <statistics/fastdds/domain/DomainParticipantImpl.hpp>
 #include <utils/shared_memory/SharedMemWatchdog.hpp>
 #include <utils/SystemInfo.hpp>
