+++ src/cpp/rtps/reader/WriterProxy.cpp
@@ -22,7 +22,7 @@
 #include <fastdds/dds/log/Log.hpp>
 #include <fastdds/rtps/writer/RTPSWriter.hpp>
 
-#include "rtps/RTPSDomainImpl.hpp"
+#include "rtps/domain/RTPSDomainImpl.hpp"
 #include "utils/collections/node_size_helpers.hpp"
 #include <rtps/builtin/data/WriterProxyData.hpp>
 #include <rtps/messages/RTPSMessageCreator.hpp>
