+++ test/mock/dds/DomainParticipantFactory/fastdds/dds/domain/DomainParticipantFactory.hpp
@@ -22,7 +22,7 @@
 
 #include <memory>
 
-#include <rtps/RTPSDomainImpl.hpp>
+#include <rtps/domain/RTPSDomainImpl.hpp>
 
 namespace eprosima {
 namespace fastdds {
